engine/pyproject.toml
[project]
name = "agent-control-engine"
version = "0.1.0"
description = "Control execution engine for Agent Control"
requires-python = ">=3.12"
dependencies = [
    "agent-control-models>=0.1.0",
    "google-re2>=1.1",
]
authors = [
    {name = "Agent Control Team"}
]
readme = "README.md"
license = {text = "Apache-2.0"}

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_control_engine"]
